+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ui
 
+## 1.5.14-next.0
+
+### Patch Changes
+
+- 0061f65: - feat: allows dev mode for cloud onboarding flow
+- Updated dependencies [0061f65]
+  - @copilotkit/shared@1.5.14-next.0
+  - @copilotkit/runtime-client-gql@1.5.14-next.0
+
 ## 1.5.13
 
 ### Patch Changes